rename dequeue directory and files
